Hello.
Here is a piece of arch/unix/fcgid_pm_unix.c, from line 286:
---------------------
   if ((rv = apr_dir_make_recursive(get_socketpath(main_server),
                                    APR_UREAD | APR_UWRITE | APR_UEXECUTE,
                                    configpool)) != APR_SUCCESS
       || chown(get_socketpath(main_server), unixd_config.user_id,
                -1) < 0) {
       ap_log_error(APLOG_MARK, APLOG_ERR, rv, main_server,
                    "mod_fcgid: Can't create unix socket dir");
       exit(1);
   }
--------------------

directory is chown()-ed even if it already exists and has
proper owner. There is no problem with it usually, as this
code is executed with root priveleges.

But, there are new solutions like SELinux where some actions
are denied even for root. For example, fedora core 5 with
default permissions allows http to bind a socket in special
directories under /var/lib/httpd, but disallow to call
chown() on the directory. As a result, the apache start
failed.

I propose to insert stat() here, which checks that the
directory already exists and have proper owner and
permissions. In that case, the chown()'ing or mkdir()'ing
are not required, and can be denied by policy.

--
Max


git-svn-id: https://svn.apache.org/repos/asf/httpd/mod_fcgid/trunk@753553 13f79535-47bb-0310-9956-ffa450edef68