The on the way up the spawn control counter increases with the process starts and terminations, in other words with site load.
On the way down the spawn control counter is decreased by time elapsed.

With the up count based on usage and the down count based on time, a
difficult situation is presented for heavily loaded sites. The up count
will always go up faster than down count if you have terminations. This is
especially true for sites that use mod_fcgid for PHP and have more than 500
req/sec. The standard is to exit PHP after 500 requests, so if you are
serving more than 500 requests per second, you bleed up on the spawn count
without bounds (unless termination score is set to zero). This bleed up
means you will eventually hit your spawn limit and new spawns will be
limited by the passage of time and not the capacity of your server.

This patch allows the user to configure a scale factor for decreasing the
spawn counter with the passage of time. For sites with large capacity
this patch allows the administrator to scale down at a rate greater than
one per second.

--tim


git-svn-id: https://svn.apache.org/repos/asf/httpd/mod_fcgid/trunk@753574 13f79535-47bb-0310-9956-ffa450edef68