Fix basic implementation of MaxRequestInMem and MaxRequestLen directives.
Previously, the size of a single input brigade (up to HUGE_STRING_LEN),
instead of the amount of request body so far, was compared with the
values of the directives.

Strangely enough, this issue was present in the original commit of
these features (r753575).

Also, improve usefulness of the warning for exceeding MaxRequestLen.


git-svn-id: https://svn.apache.org/repos/asf/httpd/mod_fcgid/trunk@816309 13f79535-47bb-0310-9956-ffa450edef68